cast size_t to int to avoid warning -- Florian octo Forster
git-svn-id: svn://svn.oetiker.ch/rrdtool/trunk/program@1317 a5681a0c-68f1-0310-ab6d-d61299d08faa

+++ src/rrd_graph.c
@@ -4084,7 +4084,7 @@ void rrd_graph_options(
                         if (size > 0) {
                             im->text_prop[propidx].size = size;
                         }
-                        if (strlen(prop) > end) {
+                        if ((int) strlen(prop) > end) {
                             if (prop[end] == ':') {
                                 strncpy(im->text_prop[propidx].font,
                                         prop + end + 1, 255);
